Turn MessageHistory and MessageRetrieveAction into interfaces
Currently MessageChannel's #getHistory methods return classes instead of
interfaces. This makes mocking impossible.

All documentation and functionality is the same as it used to be, with
only the changes that were caused by extracting the interface.

Unfortunately I have to push this to a branch with /v4 as upstream
instead of /development because only /v4 contains the other interfaces I
need. :c

Testing done: `build`